requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
dnspython==1.16.0
frozenlist==1.4.1
idna==3.6
keyboard==0.13.5
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.26.3
pandas==2.2.0
prompt-toolkit==3.0.43
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pymongo==3.11.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.2
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
yarl==1.9.4
